requirements.txt
requests==2.31.0
asterisk-ami==0.1.5
python-dotenv==1.0.0 